Easter Candy Sales
The program asks the user to enter the number of each type of easter chocolate sold, and it will create a chart with all the candy names and their sales. It will then return the sum total of the easter candies sold, and it will return the names of the candies with the highest and the lowest sales.